qallm/requirements.txt
polyglot
transformers
pandas
lingua-py
faiss-gpu
faiss-cpu